--- global.css
@@ -1,6 +1,0 @@
-body,
-body * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
