+++ index.html
@@ -7,6 +7,6 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
   </head>
   <body>
-    <script type="module" src="/index.js"></script>
+    <script type="module" src="/user-land.js"></script>
   </body>
 </html>
